restful http adapter: have any `:default_options` be placed in the query string
Currently, the adapter is smart and detects if the type of request
you are submitting has a post body (:post, :put) it will place the
options in the post body and the request doesn't (:get, :delete)
places them in the query string.

This changes it to always place any `:default_options` in the query
string, regardless of the type of request.

The reasoning for this change is that it is inconsistant where in
the request your options are depending on the type of request you
are submitting.  This makes it difficult to programmatically inspect
the request when the thing doing the inspecting either doesn't have
access to the post body (ie HAProxy for example) or the framework you
are using doesn't combine the post body params with the query params
into a single params object (like Rails does).